replace SHA1 in prod code
